% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageAggregatedSynchrony.R
\name{averageAggregatedSynchrony}
\alias{averageAggregatedSynchrony}
\title{A function to compute an aggregate (parameter 'fun2') of a function (parameter 'fun1') of
the synchrony values per time points.}
\usage{
averageAggregatedSynchrony(x, fun1 = min, fun2 = mean)
}
\value{
a number between 0 and 1 with higher values
}
\description{
A function to compute an aggregate (parameter 'fun2') of a function (parameter 'fun1') of
the synchrony values per time points.
}
