% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListOfSynchronisedBlocks.R
\name{getListOfSynchronisedBlocks}
\alias{getListOfSynchronisedBlocks}
\title{Get list of synchronised blocks}
\usage{
getListOfSynchronisedBlocks(x)
}
\description{
Get list of synchronised blocks
}
