% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdc_synchrony.R
\name{pdcsync}
\alias{pdcsync}
\title{pdc_synchrony}
\usage{
pdcsync(
  t1,
  t2,
  segment_width = 100,
  search_width = 25,
  lag_threshold = NULL,
  m = NULL,
  t = NULL
)
}
\arguments{
\item{t1}{first time series; must be a numeric vector}

\item{t2}{second time series; must be a numeric vector}

\item{segment_width}{width of the segment}

\item{search_width}{search width}
}
\description{
pdc_synchrony
}
