% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingBasedThreshold.R
\name{samplingBasedThreshold}
\alias{samplingBasedThreshold}
\title{Heuristically determine a threshold for synchronisation}
\usage{
samplingBasedThreshold(
  t1,
  t2,
  window_size,
  m,
  t,
  num_samples = 5000,
  alpha = 0.01
)
}
\arguments{
\item{window_size}{integer; size of the window}

\item{m}{embedding dimension}

\item{t}{time-delayed embedding}

\item{num_samples}{integer; number of samples to determine empirical distribution}

\item{alpha}{numeric; alpha level to determine threshold based on empirical null distribution}
}
\description{
Heuristically determine a threshold for synchronisation
}
